bot/requirements.txt
python-telegram-bot==20.7
Pillow==10.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
ffmpeg-python==0.2.0
piexif==1.1.3
